requirements.txt
Flask==2.3.2
opencv-python
numpy
pillow
pandas
spacy
transformers
torch
google-cloud-vision
